WIP remove pinia store and move it ot defaultLayout to reduce number of files
