+++ deploy/roles/frontend/templates/frontend.env
@@ -1,3 +1,2 @@
 VITE_BASE_API_URL='https://{{ main_hostname }}/api'
-VITE_SLACK_WEBHOOK_URL = {{ }}
 ENVIRONMENT='{{ environment_name }}'
